Finished background subtraction using an eigenvector based method. The function eigenback.m implements all the steps in the process and returns the query image with the background removed.
The new function foreground_mask.m is used by the eigenbackground
algorithm. It creates a binary mask indicating which pixels are
in the foreground vs background. I created this as a separate
function in case we want to use just this mask in our analysis.

Updated the examples.m file to show how to use the background
subtraction functions.

I should probably remove background_subtraction.m since it was
just a place to test out some simpler background subtraction approaches.